+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Notable changes to the `alacritty_terminal` crate are documented in its
 
 - Spurious "Failed to set new owner of XCB selection" warnings on X11
 - Lacking permissions to launch software sending Apple events
+- Mouse hints bound to a modifier also requiring shift while an application captures the mouse
 
 ## 0.17.0